+++ ingest/ecor/pipeline/styling.mplstyle
