+++ ingest/ecor/runner.py
@@ -1,11 +1,11 @@
-from ingest.ecor_richland import Pipeline
+from ingest.ecor import Pipeline
 from utils import expand, set_dev_env
 
 
 if __name__ == "__main__":
     set_dev_env()
     pipeline = Pipeline(
-        expand("config/pipeline_config_richland.yml", __file__),
+        expand("config/pipeline_config.yml", __file__),
         expand("config/storage_config.yml", __file__),
     )
     pipeline.run(
