+++ ingest/ecor/tests/data/input/sgpecorE14.00.20190910.010201.raw.2019_0910_0000_14.flx
