+++ ingest/ecor/tests/test_pipeline.py
@@ -1,16 +1,16 @@
 import os
 import xarray as xr
 from utils import expand, set_dev_env
-from ingest.ecor_richland import Pipeline
+from ingest.ecor import Pipeline
 
 parent = os.path.dirname(__file__)
 
 
 # TODO – Developer: Update paths to your input files here.
-def test_pipeline_at_richland():
+def test_pipeline():
     set_dev_env()
     pipeline = Pipeline(
-        expand("config/pipeline_config_richland.yml", parent),
+        expand("config/pipeline_config_awaken.yml", parent),
         expand("config/storage_config.yml", parent),
     )
     output = pipeline.run(expand("tests/data/input/data.csv", parent))
